requirements.txt
asgiref==3.7.2
boto3==1.28.26
botocore==1.31.26
cffi==1.15.1
cryptography==41.0.3
Django==4.2.4
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
install==1.3.5
jmespath==1.0.1
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
